requirements.txt
Django==1.6.5
Pillow==2.4.0
django-compressor==1.3
sorl-thumbnail==11.12.1b
South==0.8.4
pysolr==3.2.0
purl==0.8
django-haystack==2.1.0
beautifulsoup4==4.3.2